get all selected stats in one query
