+++ modules/scaleio/collect_system.go
@@ -5,184 +5,17 @@ import (
 	mtx "github.com/netdata/go.d.plugin/pkg/metrics"
 )
 
-/*
-Starting from version 3 of ScaleIO/VxFlex API numOfScsiInitiators property is removed from selectedSystemStatsQuery.
-Reference: VxFlex OS v3.x REST API Reference Guide.pdf
-*/
-var (
-	selectedSystemStatsQuery = `
-{
-    "selectedStatisticsList":[
-        {
-            "type":"System",
-            "properties":[
-                "bckRebuildReadBwc",
-                "bckRebuildWriteBwc",
-                "fwdRebuildReadBwc",
-                "fwdRebuildWriteBwc",
-                "normRebuildReadBwc",
-                "normRebuildWriteBwc",
-                "rebalanceReadBwc",
-                "rebalanceWriteBwc",
-                "primaryReadBwc",
-                "primaryWriteBwc",
-                "secondaryReadBwc",
-                "secondaryWriteBwc",
-                "userDataReadBwc",
-                "userDataWriteBwc",
-                "totalReadBwc",
-                "totalWriteBwc",
-                "activeBckRebuildCapacityInKb",
-                "activeFwdRebuildCapacityInKb",
-                "activeMovingCapacityInKb",
-                "activeNormRebuildCapacityInKb",
-                "activeRebalanceCapacityInKb",
-                "atRestCapacityInKb",
-                "bckRebuildCapacityInKb",
-                "capacityAvailableForVolumeAllocationInKb",
-                "capacityInUseInKb",
-                "capacityLimitInKb",
-                "degradedFailedCapacityInKb",
-                "degradedHealthyCapacityInKb",
-                "failedCapacityInKb",
-                "fwdRebuildCapacityInKb",
-                "inMaintenanceCapacityInKb",
-                "maxCapacityInKb",
-                "movingCapacityInKb",
-                "normRebuildCapacityInKb",
-                "pendingBckRebuildCapacityInKb",
-                "pendingFwdRebuildCapacityInKb",
-                "pendingMovingCapacityInKb",
-                "pendingNormRebuildCapacityInKb",
-                "pendingRebalanceCapacityInKb",
-                "protectedCapacityInKb",
-                "rebalanceCapacityInKb",
-                "semiProtectedCapacityInKb",
-                "snapCapacityInUseInKb",
-                "snapCapacityInUseOccupiedInKb",
-                "spareCapacityInKb",
-                "thickCapacityInUseInKb",
-                "thinCapacityAllocatedInKb",
-                "thinCapacityInUseInKb",
-                "unreachableUnusedCapacityInKb",
-                "unusedCapacityInKb",
-                "numOfDevices",
-                "numOfFaultSets",
-                "numOfMappedToAllVolumes",
-                "numOfProtectionDomains",
-                "numOfRfcacheDevices",
-                "numOfSdc",
-                "numOfSds",
-                "numOfSnapshots",
-                "numOfStoragePools",
-                "numOfThickBaseVolumes",
-                "numOfThinBaseVolumes",
-                "numOfUnmappedVolumes",
-                "numOfVolumes",
-                "numOfVolumesInDeletion",
-                "numOfVtrees"
-            ]
-        }
-    ]
-}
-`
-)
-
-type selectedStatistics struct {
-	System struct {
-		NumOfDevices                             float64
-		NumOfFaultSets                           float64
-		NumOfMappedToAllVolumes                  float64
-		NumOfProtectionDomains                   float64
-		NumOfRfcacheDevices                      float64
-		NumOfScsiInitiators                      float64
-		NumOfSdc                                 float64
-		NumOfSds                                 float64
-		NumOfSnapshots                           float64
-		NumOfStoragePools                        float64
-		NumOfThickBaseVolumes                    float64
-		NumOfThinBaseVolumes                     float64
-		NumOfUnmappedVolumes                     float64
-		NumOfVolumes                             float64
-		NumOfVolumesInDeletion                   float64
-		NumOfVtrees                              float64
-		ActiveBckRebuildCapacityInKb             float64
-		ActiveFwdRebuildCapacityInKb             float64
-		ActiveMovingCapacityInKb                 float64
-		ActiveNormRebuildCapacityInKb            float64
-		ActiveRebalanceCapacityInKb              float64
-		AtRestCapacityInKb                       float64
-		BckRebuildCapacityInKb                   float64
-		CapacityAvailableForVolumeAllocationInKb float64
-		CapacityInUseInKb                        float64
-		CapacityLimitInKb                        float64
-		DegradedFailedCapacityInKb               float64
-		DegradedHealthyCapacityInKb              float64
-		FailedCapacityInKb                       float64
-		FwdRebuildCapacityInKb                   float64
-		InMaintenanceCapacityInKb                float64
-		MaxCapacityInKb                          float64
-		MovingCapacityInKb                       float64
-		NormRebuildCapacityInKb                  float64
-		PendingBckRebuildCapacityInKb            float64
-		PendingFwdRebuildCapacityInKb            float64
-		PendingMovingCapacityInKb                float64
-		PendingNormRebuildCapacityInKb           float64
-		PendingRebalanceCapacityInKb             float64
-		ProtectedCapacityInKb                    float64
-		RebalanceCapacityInKb                    float64
-		SemiProtectedCapacityInKb                float64
-		SnapCapacityInUseInKb                    float64
-		SnapCapacityInUseOccupiedInKb            float64
-		SpareCapacityInKb                        float64
-		ThickCapacityInUseInKb                   float64
-		ThinCapacityAllocatedInKb                float64
-		ThinCapacityInUseInKb                    float64
-		UnreachableUnusedCapacityInKb            float64
-		UnusedCapacityInKb                       float64
-		NormRebuildReadBwc                       client.Bwc // TODO: ???
-		NormRebuildWriteBwc                      client.Bwc // TODO: ???
-		BckRebuildReadBwc                        client.Bwc // failed node/disk is back alive
-		BckRebuildWriteBwc                       client.Bwc // failed node/disk is back alive
-		FwdRebuildReadBwc                        client.Bwc // node/disk fails
-		FwdRebuildWriteBwc                       client.Bwc // node/disk fails
-		RebalanceReadBwc                         client.Bwc
-		RebalanceWriteBwc                        client.Bwc
-		PrimaryReadBwc                           client.Bwc // Backend (SDSs + Devices)
-		PrimaryWriteBwc                          client.Bwc // Backend (SDSs + Devices)
-		SecondaryReadBwc                         client.Bwc // Backend (SDSs + Devices, 2nd)
-		SecondaryWriteBwc                        client.Bwc // Backend (SDSs + Devices, 2nd)
-		UserDataReadBwc                          client.Bwc // Frontend (Volumes + SDCs)
-		UserDataWriteBwc                         client.Bwc // Frontend (Volumes + SDCs)
-		PrimaryReadFromDevBwc                    client.Bwc // TODO: ???
-		PrimaryReadFromRmcacheBwc                client.Bwc // TODO: ???
-		SecondaryReadFromDevBwc                  client.Bwc // TODO: ???
-		SecondaryReadFromRmcacheBwc              client.Bwc // TODO: ???
-		TotalReadBwc                             client.Bwc // *ReadBwc
-		TotalWriteBwc                            client.Bwc // *WriteBwc
-	}
-}
-
-func (s *ScaleIO) collectSystemOverview(mx *metrics) error {
-	var stats selectedStatistics
-	err := s.apiClient.SelectedStatistics(&stats, selectedSystemStatsQuery)
-	if err != nil {
-		return err
-	}
-
+func (s *ScaleIO) collectSystemOverview(mx *metrics, stats selectedStatistics) {
 	collectSystemCapacity(mx, stats)
 	collectSystemComponents(mx, stats)
 	collectSystemIOWorkload(mx, stats)
 	collectSystemRebalance(mx, stats)
 	collectSystemRebuild(mx, stats)
-	return nil
 }
 
 func collectSystemCapacity(mx *metrics, stats selectedStatistics) {
-	var (
-		m = &mx.SystemOverview.Capacity
-		s = stats.System
-	)
+	m := &mx.SystemOverview.Capacity
+	s := stats.System
 
 	// Physical Capacity Calculation (as in the ScaleIO GUI)
 	{
@@ -229,10 +62,8 @@ func collectSystemCapacity(mx *metrics, stats selectedStatistics) {
 }
 
 func collectSystemComponents(mx *metrics, stats selectedStatistics) {
-	var (
-		m = &mx.SystemOverview.Components
-		s = stats.System
-	)
+	m := &mx.SystemOverview.Components
+	s := stats.System
 
 	m.Devices.Set(s.NumOfDevices)
 	m.FaultSets.Set(s.NumOfFaultSets)
@@ -255,12 +86,9 @@ func collectSystemComponents(mx *metrics, stats selectedStatistics) {
 }
 
 func collectSystemIOWorkload(mx *metrics, stats selectedStatistics) {
-	var (
-		m = &mx.SystemOverview.IOWorkload
-		s = stats.System
-	)
+	m := &mx.SystemOverview.IOWorkload
+	s := stats.System
 
-	// --Bandwidth--
 	m.Total.BW.set(
 		calcBW(s.TotalReadBwc),
 		calcBW(s.TotalWriteBwc),
@@ -282,7 +110,6 @@ func collectSystemIOWorkload(mx *metrics, stats selectedStatistics) {
 		sumGauge(m.Backend.Primary.BW.Write, m.Backend.Secondary.BW.Write),
 	)
 
-	// --IOPS--
 	m.Total.IOPS.set(
 		calcIOPS(s.TotalReadBwc),
 		calcIOPS(s.TotalWriteBwc),
@@ -304,7 +131,6 @@ func collectSystemIOWorkload(mx *metrics, stats selectedStatistics) {
 		sumGauge(m.Backend.Primary.IOPS.Write, m.Backend.Secondary.IOPS.Write),
 	)
 
-	// --I/O SIZE--
 	m.Total.IOSize.set(
 		calcIOSize(s.TotalReadBwc),
 		calcIOSize(s.TotalWriteBwc),
@@ -328,12 +154,9 @@ func collectSystemIOWorkload(mx *metrics, stats selectedStatistics) {
 }
 
 func collectSystemRebuild(mx *metrics, stats selectedStatistics) {
-	var (
-		m = &mx.SystemOverview.Rebuild
-		s = stats.System
-	)
+	m := &mx.SystemOverview.Rebuild
+	s := stats.System
 
-	// --Bandwidth--
 	m.Forward.BW.set(
 		calcBW(s.FwdRebuildReadBwc),
 		calcBW(s.FwdRebuildWriteBwc),
@@ -351,7 +174,6 @@ func collectSystemRebuild(mx *metrics, stats selectedStatistics) {
 		sumGauge(m.Forward.BW.Write, m.Backward.BW.Write, m.Normal.BW.Write),
 	)
 
-	// --IOPS--
 	m.Forward.IOPS.set(
 		calcIOPS(s.FwdRebuildReadBwc),
 		calcIOPS(s.FwdRebuildWriteBwc),
@@ -369,7 +191,6 @@ func collectSystemRebuild(mx *metrics, stats selectedStatistics) {
 		sumGauge(m.Forward.IOPS.Write, m.Backward.IOPS.Write, m.Normal.IOPS.Write),
 	)
 
-	// --IO SIZE--
 	m.Forward.IOSize.set(
 		calcIOSize(s.FwdRebuildReadBwc),
 		calcIOSize(s.FwdRebuildWriteBwc),
@@ -395,30 +216,24 @@ func collectSystemRebuild(mx *metrics, stats selectedStatistics) {
 }
 
 func collectSystemRebalance(mx *metrics, stats selectedStatistics) {
-	var (
-		m = &mx.SystemOverview.Rebalance
-		s = stats.System
-	)
+	m := &mx.SystemOverview.Rebalance
+	s := stats.System
 
-	// --Bandwidth--
 	m.BW.set(
 		calcBW(s.RebalanceReadBwc),
 		calcBW(s.RebalanceWriteBwc),
 	)
 
-	// --IOPS--
 	m.IOPS.set(
 		calcIOPS(s.RebalanceReadBwc),
 		calcIOPS(s.RebalanceWriteBwc),
 	)
 
-	// --IO SIZE--
 	m.IOSize.set(
 		calcIOSize(s.RebalanceReadBwc),
 		calcIOSize(s.RebalanceWriteBwc),
 	)
 
-	// --Pending Capacity--
 	m.Pending.Set(s.PendingRebalanceCapacityInKb)
 }
